+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CDTDatastore CHANGELOG
 
+## Unreleased
+- [REMOVED] Removed deprecated class `CDTSavedHTTPAttachment` and method `createRevisionFromJson` on `CDTDocumentRevision`.
+
 ## 1.2.2 (2017-09-06)
 - [IMPROVED] Added pre-emptive session renewal when within 5 minutes of expiry.
 - [FIXED] Issue where cookie interceptor did not check for 403 HTTP response case.